Always load the up-to-date version of data.json.
When the player requests `/data.json` over HTTP, rather than return the version from the file, serialize the up-to-data data from the Campaign and send to the player.

This resolves the case in which the campaign data got modified after it was loaded on the master, but before the player has connected.