Draw "veiled" areas, hidden from players.
Veiled areas can be covered or not. For players covered areas are filled with black, and non-covered are transparent.

For master there are two modes: in normal mode covered areas are darkened and non-covered are transparent. When master presses Ctrl or Cmd on Mac, all areas become visible with covered more darkened than non-covered.

TODO: Allow master to toggle covered state for veils.