Rewrite UI framework to allow easier creation of the widget hierarchy.
Now UI can be created in a declarative way. Views are initially created with dummy panes, and then attached to real panes. Panes have `min_width`, `min_height`, `flex_width`, `flex_height` attributes that contreol how they are are embedded in a StackLayout.

Added a top-level RootLayout that listens to events from the window, with a single child.