Implement observable class attributes.
An observable attribute is defined as follows:

```python
class A(object):
    field = Attribute('field_')

    def __init__(self, value):
        self.field_ = make_observable(value)
```

Explicit deinition of `self.field_` in the constructor is optional. It will be initilized with `Observable(None)` by default.

The attribute behaves as normal class field, i.e. `a.field` returns the value and `a.field = ...` sets the value, except that setting the value will invoke the observers.